feat(build-helm-chart-release.yaml): add step to commit latest release version
Add a new step to commit the latest release version. This step sets the git user name and email to "GitHub Action" and "action@github.com" respectively. It then adds the updated Chart.yaml file to the commit and commits it with the message "Update tags in YAML files". Finally, it pushes the commit to the repository.