feat(workflows): add git commit and push for updated helm chart version and code-server version
This commit adds a new step to the build-helm-chart-release.yaml workflow
that gets the latest release version of the helm chart and commits it to
the repository. It also adds a new step to the check_ha_release.yml
workflow that gets the latest release version of code-server and commits
it to the repository.